requirements.txt
python-telegram-bot[webhooks]==21.10
httpx==0.28.1
beautifulsoup4==4.13.4
python-dotenv==1.1.0
APScheduler==3.10.4
tenacity==9.1.2
pytest==8.3.5
SQLAlchemy==2.0.41
psycopg[binary]==3.2.9
